iam_portal/backend/requirements.txt
fastapi
uvicorn
google-cloud-firestore
pydantic
pydantic-settings
python-multipart
